PPU: Fixes for 16x32 & 32x64 sprites and tile count/flip
With the undocumented sizes 16x32 and 32x64 two of the obj size signals
are active simultaneously. It seems that the Range and Time periods use
a mix of the two sizes for different checks.

Example: With a sprite size of 16x32 the sprite is still in range when
the H position is set to -16. Time treats the sprite as 16x16 but with H
position -16 it displays the tiles with horizontal offset 2 & 3 instead
of 0 & 1.

Normally with sprite interlace enabled, the sprite height is halved but
with a sprite size of 16x32 it is treated as 32x8 instead of 32x16.

With tile flip it only inverts the tile count bits according to sprite size.